Refactor : 모달 컴포넌트 위치 변경
